+++ src/main/java/com/woosan/hr_system/vacation/service/VacationService.java
@@ -13,4 +13,5 @@ public interface VacationService {
     String requestVacation(Vacation vacation);
     String editVacation(Vacation vacation);
     String processVacation(int vacationId, String status);
+    String deleteVacation(int vacationId);
 }
