+++ src/main/resources/static/css/vacation/detail.css
@@ -1,0 +1,77 @@
+.vacation-form-container {
+    line-height: 1.6;
+    width: 210mm;
+    margin: 0 auto;
+    padding: 15mm;
+    box-sizing: border-box;
+    border: 1px solid #000;
+    font-family: Arial, sans-serif;
+}
+
+.btn-container {
+    display: flex;
+    justify-content: end;
+    gap: 10px;
+}
+
+textarea {
+    width: 550px;
+    height: 200px;
+    font-size: 16px;
+    resize: none;
+}
+
+.title {
+    text-align: center;
+    font-size: 24px;
+    font-weight: bold;
+    margin-bottom: 30px;
+}
+
+.section-title {
+    font-weight: bold;
+    text-align: center;
+    background-color: #f0f0f0;
+}
+
+.details, .employee-details {
+    width: 100%;
+    border-collapse: collapse;
+    margin-bottom: 10px;
+}
+
+.details td, .employee-details td {
+    border: 1px solid #000;
+    padding: 8px;
+    white-space: nowrap;
+    text-align: center;
+}
+
+.submission-declaration {
+    text-align: center;
+    font-weight: bold;
+    font-size: 20px;
+}
+
+.applicant-signature {
+    display: flex;
+    flex-direction: column;
+    gap: 5px;
+    align-items: end;
+    font-size: 17px;
+}
+
+.company-signature {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+}
+.company-signature p {
+    font-size: 30px;
+}
+.signature {
+    margin-left: 5px;
+    width: 50px;
+    height: 50px;
+}
