+++ src/main/resources/templates/vacation/employee.html
@@ -4,6 +4,7 @@
       layout:decorate="~{fragments/layout-with-modal :: layout(title='휴가 내역 조회')}">
 
 <th:block layout:fragment="css">
+    <link rel="stylesheet" th:href="@{/css/vacation/detail.css}"/>
     <link rel="stylesheet" th:href="@{/css/pagination.css}"/>
     <style>
         .list tbody tr:hover {
@@ -25,6 +26,7 @@
 </th:block>
 
 <th:block layout:fragment="script">
+    <script th:src="@{/js/vacation/ajax.js}"></script>
     <script th:src="@{/js/common/format.js}"></script>
     <script>
         document.addEventListener("DOMContentLoaded", function() {
